Initial inkpot source
This is the file as it may be seen at https://github.com/ciaranm/inkpot

I made this new repository because inkpot was left unmaintained for the
last 9 years.
Today, new features came both in vim and in terminals, but the current
inkpot theme doesn't support them. This is sad because adding support
for theses features would be quite simple in some cases.

My goal with this repository is to update the inkpot colorscheme, not
its palette, and make it catch up with vim and neovim.