Add extra highlight groups for each mode
New highlight groups:
  NormalMode
  InsertMode
  CommandMode
  VisualMode
  ReplaceMode

These are mainly for personal use in my statusline.
This is not my idea, this is @lifepillar's and their colorscheme
(solarized8 and more) have theses undocumented highlight groups, hidden
among the other ones.

You can think of it as a replacement for vim-airline's or
vim-lightline's highlights (which I may add)