Change theme screenshot
I change the screenshot because of licensing issues with the previous
one, as it was showing the SF Mono font, which is a proprietary Apple
font, and its license do not allow to publish screenshot of software
running on other platforms than macOS, iOS et al. showing this font.

As you may have guessed, this is *not* macOS, and because I don't want
to get sued (though who would sue me for a screenshot of a colorscheme ?),
I need to change the screenshot.

As I write this commit, I realise that I have kind of taken an
unmaintained project, to refactor it, and leave it unmaintained after
that.
That wasn't my initial goal...